Add files via upload
This is a python file that uses PySpark to generate word frequencies in the story body of all stories. We can apply this to our analysis because it extracts the characters who are most often included in the stories on AO3. In this file, the fandoms are isolated by tv shows, but the analysis could potentially be further improved by including all fandom crossovers and analyzing word frequencies for popular stories only (i.e. hits or kudos above 3 standard deviations).
The Good Place
+-------+-----+                                                                 
|   word|count|
+-------+-----+
|eleanor|28378|
| tahani|15993|
|michael|11243|
|  chidi|10258|
|  janet| 7294|
|   good| 5672|
|  jason| 4522|
|  place| 4489|
|  right| 4118|
| really| 3943|
+-------+-----+
only showing top 10 rows

Brooklyn Nine-Nine
+------+------+                                                                 
|  word| count|
+------+------+
|  jake|152896|
|   amy|129926|
|  rosa| 53991|
|  gina| 34315|
|  face| 21875|
|   way| 21799|
|  holt| 21315|
|  hand| 21130|
| right| 21020|
|around| 20957|
+------+------+
only showing top 10 rows

Parks and Rec
+------+-----+                                                                  
|  word|count|
+------+-----+
|leslie|50671|
|   ben|43486|
|   ann|20237|
| april|17220|
|  andy|12778|
|really| 9947|
|  hand| 8884|
|around| 8794|
| right| 8577|
|   way| 8097|
+------+-----+
only showing top 10 rows

The Office
+-------+-----+                                                                 
|   word|count|
+-------+-----+
|    jim|42611|
|    pam|27166|
| dwight|22659|
|michael|11593|
|   ryan|10190|
| really| 7119|
|   andy| 6752|
|   head| 6378|
|   hand| 6285|
| around| 6191|
+-------+-----+